pipeline/requirements.txt
# Cloud Function dependencies
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-documentai==2.*
google-auth==2.*
requests==2.*
